[Feat]: refactor styles into button.styles file and add hover and focus styles to the solid and outline variants
