--- src/components/ui/button/button.module.scss
@@ -1,40 +1,0 @@
-.button {
-  color: var(--color-text);
-}
-
-.button__text {
-  color: var(--color-text);
-}
-.disabled__text {
-  color: var(--color-text) !important;
-  opacity: 0.5;
-}
-
-.button__outlined {
-  border-color: var(--color-border);
-
-  &:hover {
-    border-color: var(--color-border);
-    opacity: 0.9;
-  }
-}
-
-.disabled__outlined {
-  color: var(--color-text) !important;
-  border-color: var(--color-border) !important;
-  opacity: 0.5;
-}
-
-.button__contained {
-  color: var(--color-white-black);
-  background-color: var(--color-black-white);
-  &:hover {
-    background-color: var(--color-black-white);
-  }
-}
-
-.disabled__contained {
-  color: var(--color-white-black) !important;
-  background-color: var(--color-black-white) !important;
-  opacity: 0.5;
-}
